refactor: :building_construction: creates scopes for pages
